repair

+++ Bachelor_app/Extension/MatExtension.cs
@@ -102,7 +102,7 @@ namespace Bachelor_app.Extension
                 case DepthType.Cv16S: return new short[1];
                 case DepthType.Cv16U: return new ushort[1];
                 case DepthType.Cv32S: return new int[1];
-                case DepthType.Cv32F: return new sbyte[1];
+                case DepthType.Cv32F: return new float[1];
                 case DepthType.Cv64F: return new double[1];
                 default: return new float[1];
             }
